requirements.txt
albumentations==1.4.21
beautifulsoup4==4.12.3
faiss_cpu==1.9.0
Flask==3.1.0
Flask_Login==0.6.3
flask_mail==0.10.0
gunicorn
imgaug==0.4.0
matplotlib==3.10.0
opencv-python-headless==4.9.0.80
opencv-contrib-python-headless==4.9.0.80
numpy<2
pdfplumber==0.11.4
python-dotenv==1.0.1
Requests==2.32.3
seaborn==0.13.2
sendgrid==6.11.0
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
Werkzeug==3.1.3
yt_dlp==2024.12.23
